+++ docs/changelog.rst
@@ -49,7 +49,7 @@ Changelog
   no longer set to ``0``.
 - :gh:`2977`: new :func:`bytes2human` utility function, converting a number of
   bytes to a human-readable string (e.g. ``9.8K``).
-- :gh:`1452`, [Linux]: :meth:`Process.memory_footprint` returns a new
+- :gh:`1452`, [Linux, Windows]: :meth:`Process.memory_footprint` returns a new
   :field:`shared` field: resident memory actually shared with other processes.
 
 Reorganization of process memory APIs (:gh:`2731`, :gh:`2736`, :gh:`2723`,